+++ angular-workspace/projects/ni/spright-angular/CHANGELOG.md
@@ -1,8 +1,33 @@
 # Change Log - @ni/spright-angular
 
-This log was last generated on Wed, 24 Apr 2024 18:22:05 GMT and should not be manually modified.
+This log was last generated on Fri, 26 Apr 2024 21:13:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.1.4
+
+Fri, 26 Apr 2024 21:13:31 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.7
+
+## 0.1.3
+
+Fri, 26 Apr 2024 20:31:48 GMT
+
+### Patches
+
+- No API changes, validating release pipelines ([ni/nimble@c5a74e6](https://github.com/ni/nimble/commit/c5a74e68c2909e2f9a50bf0e6b0e4f1c5f0cbce8))
+- Bump @ni/spright-components to v0.0.6
+
+## 0.1.1
+
+Thu, 25 Apr 2024 20:42:08 GMT
+
+### Patches
+
+- Bump @ni/spright-components to v0.0.4
 
 ## 0.1.0