+++ packages/nimble-blazor/NimbleBlazor/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-blazor
 
-This log was last generated on Mon, 22 Apr 2024 15:13:13 GMT and should not be manually modified.
+This log was last generated on Fri, 26 Apr 2024 20:31:47 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 16.0.3
+
+Fri, 26 Apr 2024 20:31:47 GMT
+
+### Patches
+
+- No API changes, validating release pipelines ([ni/nimble@c5a74e6](https://github.com/ni/nimble/commit/c5a74e68c2909e2f9a50bf0e6b0e4f1c5f0cbce8))
 
 ## 16.0.0